third_party/WebKit/LayoutTests/http/tests/security/suborigins/suborigin-allow-same-suborigin-access.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>Allow suborigin in HTTP header</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<script>
var tests;
function next_test() {
  var test = tests.shift();
  if (test)
    test();
}

function make_iframe_test(test_name, resource) {
  return function() {
  async_test(t => {
      window.onmessage = t.step_func(function(event) {
          assert_equals(event.data, 'I am a secret');
          // next_test() must be called before t.done() so that the next test,
          // if one exists, is setup and the overall test doesn't end.
          next_test();
          t.done();
        });
      var iframe = document.createElement('iframe');
      iframe.src = 'resources/' + resource + '?' +
                   'suborigin=foobar&childsuborigin=foobar';
      document.body.appendChild(iframe);
    }, test_name);
  }
}

var tests = [
  make_iframe_test(
    'iframe reaches into parent', 'iframe-reaches-into-parent.php'),
  make_iframe_test(
    'parent reaches into iframe', 'reach-into-iframe.php')
];

next_test();
</script>
</body>
</html>
